作成するにあたってわかりづらかったのでディレクトリ名をsettingholidayからholidayに変更しました。それに合わせてすべてのファイルが変更されています。
